+++ lib/tasks/import/centos_advisories.rake
@@ -1,6 +1,6 @@
 namespace :import do
   desc 'Import centos advisories from the web'
   task centos_advisories: :environment do
-    Import.new.centos_advisories()
+    Import::Yum.new.centos_advisories()
   end
 end
