+++ lib/tasks/import/rhel_advisories.rake
@@ -1,6 +1,6 @@
 namespace :import do
   desc 'Import rhel advisories from the web'
   task rhel_advisories: :environment do
-    Import.new.rhel_advisories()
+    Import::Yum.new.rhel_advisories()
   end
 end
