+++ lib/tasks/import/servers.rake
@@ -1,6 +1,6 @@
 namespace :import do
   desc 'Import package reports from our servers'
   task servers: :environment do
-    Import.new.servers()
+    Import::Servers.new.servers()
   end
 end
